+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-grid.pipes.ts
@@ -3,7 +3,7 @@ import { cloneArray } from '../../core/utils';
 import { DataUtil } from '../../data-operations/data-util';
 import { FilteringExpressionsTree, IFilteringExpressionsTree } from '../../data-operations/filtering-expressions-tree';
 import { IFilteringStrategy } from '../../data-operations/filtering-strategy';
-import { DEFAULT_PIVOT_KEYS, IPivotConfiguration, IPivotDimension, IPivotKeys } from './pivot-grid.interface';
+import { DEFAULT_PIVOT_KEYS, IPivotConfiguration, IPivotDimension, IPivotKeys, PivotDimensionType } from './pivot-grid.interface';
 import {
     DefaultPivotSortingStrategy, DimensionValuesFilteringStrategy, PivotColumnDimensionsStrategy,
     PivotRowDimensionsStrategy
@@ -15,6 +15,8 @@ import { GridType, IGX_GRID_BASE } from '../common/grid.interface';
 import { GridBaseAPIService } from '../api.service';
 import { IgxGridBaseDirective } from '../grid-base.directive';
 import { IGridSortingStrategy } from '../common/strategy';
+import { IGroupByResult } from '../../data-operations/grouping-result.interface';
+import { IGroupingExpression } from '../../data-operations/grouping-expression.interface';
 
 /**
  * @hidden
@@ -96,6 +98,58 @@ export class IgxPivotRowExpansionPipe implements PipeTransform {
                 }
             });
         });
+    }
+}
+
+/**
+ * @hidden
+ */
+@Pipe({
+    name: 'pivotGridCellMerging',
+    pure: true
+})
+export class IgxPivotCellMergingPipe implements PipeTransform {
+    constructor(@Inject(IGX_GRID_BASE) private grid: GridType) { }
+    public transform(
+        collection: any[],
+        config: IPivotConfiguration,
+        dim: IPivotDimension,
+        pivotKeys: IPivotKeys,
+        _pipeTrigger?: number
+    ): any[] {
+        if (collection.length === 0 || config.rows.length === 0) return collection;
+        const data = collection ? cloneArray(collection, true) : [];
+        const res = [];
+
+        const enabledRows = config.rows.filter(x => x.enabled);
+
+        const prevDims = enabledRows.filter((d, ind) => ind < enabledRows.indexOf(dim));
+        let groupData = [];
+        let prevDim;
+        let prevDimRoot;
+        let prevId;
+        for (let rec of data) {
+            const dimData = PivotUtil.getDimensionLevel(dim, rec, pivotKeys);
+            const id = PivotUtil.getRecordKey(rec, dimData.dimension, prevDims, pivotKeys);
+            if (groupData.length > 0 && prevId !== id) {
+                const h = groupData.length > 1 ? groupData.length * this.grid.renderedRowHeight : undefined;
+                groupData[0][prevDimRoot.memberName + pivotKeys.rowDimensionSeparator + 'height'] = h;
+                groupData[0][prevDim.dimension.memberName + pivotKeys.rowDimensionSeparator + 'rowSpan'] = groupData.length;
+                res.push(groupData[0]);
+                groupData = [];
+            }
+            groupData.push(rec);
+            prevDim = dimData;
+            prevDimRoot = dim;
+            prevId = id;
+        }
+        if (groupData.length > 0) {
+            const h = groupData.length > 1 ? groupData.length * this.grid.rowHeight + (groupData.length - 1) + 1 : undefined;
+            groupData[0][prevDimRoot.memberName + pivotKeys.rowDimensionSeparator + 'height'] = h;
+            groupData[0][prevDim.dimension.memberName + pivotKeys.rowDimensionSeparator + 'rowSpan'] = groupData.length;
+            res.push(groupData[0]);
+        }
+        return res;
     }
 }