+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-header-row.component.ts
@@ -4,8 +4,10 @@ import {
     Component,
     ElementRef,
     Inject,
+    OnChanges,
     QueryList,
     Renderer2,
+    SimpleChanges,
     ViewChildren
 } from '@angular/core';
 import { first } from 'rxjs/operators';
@@ -13,14 +15,17 @@ import { IBaseChipEventArgs, IgxChipComponent } from '../../chips/chip.component
 import { IgxChipsAreaComponent } from '../../chips/chips-area.component';
 import { GridColumnDataType } from '../../data-operations/data-util';
 import { SortingDirection } from '../../data-operations/sorting-strategy';
+import { IgxGridForOfDirective } from '../../directives/for-of/for_of.directive';
 import { ISelectionEventArgs } from '../../drop-down/drop-down.common';
 import { IgxDropDownComponent } from '../../drop-down/drop-down.component';
 import { AbsoluteScrollStrategy, AutoPositionStrategy, OverlaySettings, PositionSettings, VerticalAlignment } from '../../services/public_api';
-import { IGX_GRID_BASE, PivotGridType } from '../common/grid.interface';
+import { ColumnType, IGX_GRID_BASE, PivotGridType } from '../common/grid.interface';
+import { IgxGridHeaderGroupComponent } from '../headers/grid-header-group.component';
 import { IgxGridHeaderRowComponent } from '../headers/grid-header-row.component';
 import { DropPosition } from '../moving/moving.service';
 import { IgxPivotAggregate, IgxPivotDateAggregate, IgxPivotNumericAggregate, IgxPivotTimeAggregate } from './pivot-grid-aggregate';
 import { IPivotAggregator, IPivotDimension, IPivotValue, PivotDimensionType } from './pivot-grid.interface';
+import { PivotUtil } from './pivot-util';
 
 /**
  *
@@ -35,7 +40,7 @@ import { IPivotAggregator, IPivotDimension, IPivotValue, PivotDimensionType } fr
     selector: 'igx-pivot-header-row',
     templateUrl: './pivot-header-row.component.html'
 })
-export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
+export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent implements OnChanges {
     public aggregateList: IPivotAggregator[] = [];
 
     public value: IPivotValue;
@@ -61,32 +66,167 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         super(ref, cdr);
     }
 
+    /**
+    * @hidden
+    * @internal
+    * The virtualized part of the header row containing the unpinned header groups. */
+    @ViewChildren('headerVirtualContainer', { read: IgxGridForOfDirective })
+    public headerContainers: QueryList<IgxGridForOfDirective<IgxGridHeaderGroupComponent>>;
+
+    public get headerForOf() {
+        return this.headerContainers.last;
+    }
+
+    /**
+    * @hidden
+    * @internal
+    */
     @ViewChildren('notifyChip')
     public notificationChips: QueryList<IgxChipComponent>;
 
+    /**
+    * @hidden
+    * @internal
+    */
+    public columnDimensionsByLevel: any[] = [];
+
+    /**
+    * @hidden
+    * @internal
+    */
+    public get totalDepth() {
+        const columnDimensions = this.grid.columnDimensions;
+        if (columnDimensions.length === 0) {
+            return 1;
+        }
+        let totalDepth = columnDimensions.map(x => PivotUtil.getDimensionDepth(x) + 1).reduce((acc, val) => acc + val);
+        if (this.grid.hasMultipleValues) {
+            totalDepth += 1;
+        }
+        return totalDepth;
+    }
+
+    /**
+    * @hidden
+    * @internal
+    */
+    public get maxContainerHeight() {
+        return this.totalDepth > 1 ? this.totalDepth * this.grid.renderedRowHeight : undefined;
+    }
+
+    /**
+    * @hidden
+    * @internal
+    */
+    public calcHeight(col: ColumnType, index: number) {
+        return !col.columnGroup && col.level < this.totalDepth && col.level === index ? (this.totalDepth - col.level) * this.grid.rowHeight : this.grid.rowHeight;
+    }
+
+    /**
+    * @hidden
+    * @internal
+    */
+    public isDuplicateOfExistingParent(col: ColumnType, lvl: number) {
+        const parentCollection = lvl > 0 ? this.columnDimensionsByLevel[lvl - 1] : [];
+        const duplicate = parentCollection.indexOf(col) !== -1;
+
+        return duplicate;
+    }
+
+    /**
+    * @hidden
+    * @internal
+    */
+    public isMultiRow(col: ColumnType, lvl: number) {
+        const isLeaf = !col.columnGroup;
+        return isLeaf && lvl !== this.totalDepth - 1;
+    }
+
+    /**
+    * @hidden
+    * @internal
+    */
+    public populateColumnDimensionsByLevel() {
+        const res = [];
+        const columnDimensions = this.grid.columnDimensions;
+        if (columnDimensions.length === 0) {
+            this.columnDimensionsByLevel = res;
+            return;
+        }
+        for (let i = 0; i < this.totalDepth; i++) {
+            res[i] = [];
+        }
+        const cols = this.unpinnedColumnCollection;
+        // populate column dimension matrix recursively
+        this.populateDimensionRecursively(cols.filter(x => x.level === 0), 0, res);
+        this.columnDimensionsByLevel = res;
+    }
+
+    protected populateDimensionRecursively(currentLevelColumns: ColumnType[], level = 0, res: any[]) {
+        currentLevelColumns.forEach(col => {
+            if (res[level]) {
+                res[level].push(col);
+                if (col.columnGroup && col.children.length > 0) {
+                    const visibleColumns = col.children.toArray().filter(x => !x.hidden);
+                    this.populateDimensionRecursively(visibleColumns, level + 1, res);
+                } else if (level < this.totalDepth - 1) {
+                    for (let i = level + 1; i <= this.totalDepth - 1; i++) {
+                        res[i].push(col);
+                    }
+                }
+            }
+        });
+    }
+
+    /**
+    * @hidden
+    * @internal
+    */
+    public ngOnChanges(changes: SimpleChanges) {
+        if (changes.unpinnedColumnCollection && this.unpinnedColumnCollection.length > 0) {
+            this.populateColumnDimensionsByLevel();
+        }
+    }
+
+    /**
+    * @hidden
+    * @internal
+    */
     public onDimDragStart(event, area) {
         this.cdr.detectChanges();
         for (let chip of this.notificationChips) {
             if (area.chipsList.toArray().indexOf(chip) === -1 &&
-             chip.nativeElement.parentElement.children.length > 0 &&
-             chip.nativeElement.parentElement.children.item(0).id !== 'empty' ) {
+                chip.nativeElement.parentElement.children.length > 0 &&
+                chip.nativeElement.parentElement.children.item(0).id !== 'empty') {
                 chip.nativeElement.hidden = false;
                 chip.nativeElement.scrollIntoView();
             }
         }
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onDimDragEnd() {
         for (let chip of this.notificationChips) {
             chip.nativeElement.hidden = true;
         }
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public getAreaHeight(area: IgxChipsAreaComponent) {
-        const chips =  area.chipsList;
+        const chips = area.chipsList;
         return chips && chips.length > 0 ? chips.first.nativeElement.clientHeight : 0;
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public getAggregateList(val: IPivotValue): IPivotAggregator[] {
         if (!val.aggregateList) {
             let defaultAggr = this.getAggregatorsForValue(val);
@@ -106,6 +246,10 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         return val.aggregateList;
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public rowRemoved(event: IBaseChipEventArgs) {
         const row = this.grid.pivotConfiguration.rows.find(x => x.memberName === event.owner.id);
         row.enabled = false;
@@ -114,16 +258,24 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         this.grid.dimensionsChange.emit({ dimensions: this.grid.pivotConfiguration.rows, dimensionCollectionType: PivotDimensionType.Row });
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public columnRemoved(event: IBaseChipEventArgs) {
         const col = this.grid.pivotConfiguration.columns.find(x => x.memberName === event.owner.id);
         col.enabled = false;
         this.grid.setupColumns();
         this.grid.filteringService.clearFilter(col.memberName);
         this.grid.pipeTrigger++;
-        this.grid.dimensionsChange.emit({dimensions: this.grid.pivotConfiguration.columns, dimensionCollectionType: PivotDimensionType.Row});
+        this.grid.dimensionsChange.emit({ dimensions: this.grid.pivotConfiguration.columns, dimensionCollectionType: PivotDimensionType.Row });
         this.grid.reflow();
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public valueRemoved(event: IBaseChipEventArgs) {
         const value = this.grid.pivotConfiguration.values.find(x => x.member === event.owner.id || x.displayName === event.owner.id);
         value.enabled = false;
@@ -132,6 +284,10 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         this.grid.valuesChange.emit({ values: this.grid.pivotConfiguration.values });
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public filterRemoved(event: IBaseChipEventArgs) {
         const filter = this.grid.pivotConfiguration.filters.find(x => x.memberName === event.owner.id);
         filter.enabled = false;
@@ -140,11 +296,19 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         this.grid.dimensionsChange.emit({ dimensions: this.grid.pivotConfiguration.filters, dimensionCollectionType: PivotDimensionType.Filter });
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onFilteringIconPointerDown(event) {
         event.stopPropagation();
         event.preventDefault();
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onFilteringIconClick(event, dimension) {
         event.stopPropagation();
         event.preventDefault();
@@ -161,6 +325,10 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         this.grid.filteringService.toggleFilterDropdown(event.target, col);
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onSummaryClick(eventArgs, value: IPivotValue, dropdown: IgxDropDownComponent, chip: IgxChipComponent) {
         this._subMenuOverlaySettings.target = eventArgs.currentTarget;
         if (dropdown.collapsed) {
@@ -174,6 +342,10 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         }
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onAggregationChange(event: ISelectionEventArgs) {
         if (!this.isSelected(event.newSelection.value)) {
             this.value.aggregate = event.newSelection.value;
@@ -181,10 +353,18 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         }
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public isSelected(val: IPivotAggregator) {
         return this.value.aggregate.key === val.key;
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onChipSort(event, dimension: IPivotDimension, dimensionType: PivotDimensionType) {
         if (!dimension.sortDirection) {
             dimension.sortDirection = SortingDirection.None;
@@ -203,6 +383,10 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         }
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onDimDragOver(event, dimension?: PivotDimensionType) {
         const typeMismatch = dimension !== undefined ? this.grid.pivotConfiguration.values.find(x => x.member === event.dragChip.id
             || x.displayName === event.dragChip.id) :
@@ -226,12 +410,20 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         }
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onDimDragLeave(event) {
         event.owner.nativeElement.previousElementSibling.style.visibility = 'hidden';
         event.owner.nativeElement.nextElementSibling.style.visibility = 'hidden';
         this._dropPos = DropPosition.AfterDropTarget;
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onAreaDragLeave(event, area) {
         const dataChips = area.chipsList.toArray().filter(x => this.notificationChips.toArray().indexOf(x) === -1);
         dataChips.forEach(element => {
@@ -242,6 +434,10 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         });
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onValueDrop(event, area) {
         //values can only be reordered
         const currentDim = this.grid.pivotConfiguration.values;
@@ -259,6 +455,10 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
         }
     }
 
+    /**
+    * @hidden
+    * @internal
+    */
     public onDimDrop(event, area, dimension: PivotDimensionType) {
         const dragId = event.dragChip?.id || event.dragData?.chip.id;
         const currentDim = this.getDimensionsByType(dimension);
@@ -275,26 +475,19 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
                 .concat(this.grid.pivotConfiguration.columns)
                 .concat(this.grid.pivotConfiguration.filters);
             // chip moved from external collection
-            const dims = allDims.filter(x => x && x.memberName === dragId);
-            if (dims.length === 0) {
+            const dim = allDims.find(x => x && x.memberName === dragId);
+            if (!dim) {
                 // you have dragged something that is not a dimension
                 return;
             }
-            dims.forEach(element => {
-                element.enabled = false;
-            });
+            const dimType = this.getDimensionsType(dim);
 
-            const currentDimChild = currentDim.find(x => x && x.memberName === dragId);
-            if (currentDimChild) {
-                currentDimChild.enabled = true;
-                const dragChipIndex = currentDim.indexOf(currentDimChild);
-                currentDim.splice(dragChipIndex, 1);
-                currentDim.splice(dragChipIndex > chipIndex ? targetIndex : targetIndex - 1, 0, currentDimChild);
-            } else {
-                const newDim = Object.assign({}, dims[0]);
-                newDim.enabled = true;
-                currentDim.splice(targetIndex, 0, newDim);
-            }
+            // Dragged chip from a different dimension to the current one.
+            const prevDimensionCollection = this.getDimensionsByType(dimType);
+            // delete from previous dimension collection and add to current.
+            prevDimensionCollection.splice(prevDimensionCollection.indexOf(dim), 1);
+            currentDim.splice(targetIndex, 0, dim);
+
             const isDraggedFromColumn = !!this.grid.pivotConfiguration.columns?.find(x => x && x.memberName === dragId);
             if (isDraggedFromColumn) {
                 // columns have changed.
@@ -339,6 +532,12 @@ export class IgxPivotHeaderRowComponent extends IgxGridHeaderRowComponent {
             default:
                 return null;
         }
+    }
+
+    protected getDimensionsType(dimension: IPivotDimension) {
+        const isColumn = !!this.grid.pivotConfiguration.columns?.find(x => x && x.memberName === dimension.memberName);
+        const isRow = !!this.grid.pivotConfiguration.rows?.find(x => x && x.memberName === dimension.memberName);
+        return isColumn ? PivotDimensionType.Column : isRow ? PivotDimensionType.Row : PivotDimensionType.Filter;
     }
 
     protected getAggregatorsForValue(value: IPivotValue): IPivotAggregator[] {