+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-row-dimension-header.component.ts
@@ -3,7 +3,7 @@ import { ChangeDetectionStrategy, ChangeDetectorRef, Component, ElementRef, Host
 import { GridType, IGX_GRID_BASE } from '../common/grid.interface';
 
 import { IgxGridHeaderComponent } from '../headers/grid-header.component';
-import { IgxColumnResizingService } from '../resizing/resizing.service';
+import { IgxPivotColumnResizingService } from '../resizing/pivot-grid/pivot-resizing.service';
 
 /**
  * @hidden
@@ -17,9 +17,9 @@ export class IgxPivotRowDimensionHeaderComponent extends IgxGridHeaderComponent 
 
     constructor(
         @Inject(IGX_GRID_BASE) public grid: GridType,
-        public colResizingService: IgxColumnResizingService,
+        public colResizingService: IgxPivotColumnResizingService,
         public cdr: ChangeDetectorRef,
-        private refInstance: ElementRef<HTMLElement>
+        public refInstance: ElementRef<HTMLElement>
     ) {
         super(grid, colResizingService, cdr, refInstance);
     }